refactor: remove duplicate values and adjust config inheritance to local <- k8s-local <- dynamic
